Conscious Chat, AI Emulations
Start of a new project aimed at recreating the conscious interations
and having the visual, auditory cues created and produced as output
to the human.

Multi user interactions amongst multiple AI entities creating a
game of life like phenomenon to continue and play out for human
entertainment and question answering.